refactor: segregate language interfaces
