+++ src/presentation/languages/index.ts
@@ -1,2 +1,1 @@
-export * from "./ILanguage";
 export * from "./i18n";
