+++ src/presentation/languages/interfaces/ICreateHabitControllerLanguage.ts
@@ -1,0 +1,9 @@
+import { IHabitAlreadyExistsErrorLanguage } from "./IHabitAlreadyExistsErrorLanguage";
+import { IMissingParamsErrorLanguage } from "./IMissingParamsErrorLanguage";
+
+export interface ICreateHabitControllerLanguage
+  extends IHabitAlreadyExistsErrorLanguage,
+    IMissingParamsErrorLanguage {
+  getHabitCreatedSuccessfullyMessage(name: string): string;
+  getHabitNotCreatedMessage(name: string): string;
+}
