+++ src/presentation/languages/interfaces/IHabitAlreadyExistsErrorLanguage.ts
@@ -1,0 +1,3 @@
+export interface IHabitAlreadyExistsErrorLanguage {
+  getHabitAlreadyExistsErrorMessage(name: string): string;
+}
