+++ src/presentation/languages/interfaces/IIndexHabitControllerLanguage.ts
@@ -1,0 +1,4 @@
+export interface IIndexHabitControllerLanguage {
+  getHabitsLoadedSuccessfullyMessage(): string;
+  getHabitsNotLoadedMessage(): string;
+}
