+++ src/presentation/languages/interfaces/IMissingParamsErrorLanguage.ts
@@ -1,0 +1,5 @@
+export interface IMissingParamsErrorLanguage {
+  getMissingParamsErrorMessage(params: string[]): string;
+  getHabitNameParamMessage(): string;
+  getPerformedLastDateParamMessage(): string;
+}
