+++ src/presentation/languages/interfaces/IToastLanguage.ts
@@ -1,0 +1,4 @@
+export interface IToastLanguage {
+  getSuccessToastTitle(): string;
+  getErrorToastTitle(): string;
+}
